[Fix] 차트 - 미션 0개일때 예외처리

+++ app/src/main/java/com/mtjin/envmate/data/chart/ChartRepository.kt
@@ -14,12 +14,12 @@ interface ChartRepository {
     fun requestCompareIndustrySameAll(usage: Int): Single<EnvRes>
 
     fun requestDetailIndustryEnergy(
-        gas: Float,
-        other: Float,
-        oil: Float,
-        coal: Float,
-        thermal: Float,
-        electric: Float
+        gas: Int,
+        other: Int,
+        oil: Int,
+        coal: Int,
+        thermal: Int,
+        electric: Int
     ): Single<IndustryEnergyRes>
 
 }
